refactor(chrome): browser-style topbar with address bar
Polish the app chrome to match a browser UX idiom:

- Brand: "PAPILLON_SYS" → "Papillon"; butterfly logo + name flush-left
  in a 64px zone that aligns with the sidebar column below it. Clicking
  the brand opens the canvas switcher dropdown (replacing the hamburger).
- Address bar: the InlinePrompt is promoted to a full-width pill input
  in the topbar center zone, always accessible regardless of page.
  pap:// autocomplete and arrow-key navigation are preserved.
- Status: the READY badge, DID display, and settings gear are removed
  from the topbar — settings lives in the sidebar, status is shown as
  a single 8px dot (teal/gold/blue) in the right end zone.
- Canvas page: canvas-prompt-bar removed; blocks/empty-state render
  directly below the topbar with no extra gap layer.
